add: CRD member

+++ controllers/projects.js
@@ -1,25 +1,33 @@
 const models = require('../models')
 const Project = models.Project
 
-exports.projects_create = async (req, res, next) => {
-  console.log(req.usuario)
-  Project.create(req.body).then(o => {
-    res.send(o.dataValues)
-  })
+exports.projects_create = async (req, res) => {
+  Project.create(req.body).then(o => res.send(o.dataValues))
 }
 
-exports.projects_get_all = async (req, res, next) => {
-  await Project.findAll().then(o => res.send(o.dataValues))
+exports.projects_get_all = async (req, res) => {
+  const lista = await Project.findAll()
+  return res.send(lista)
 }
 
 exports.projects_get = async (req, res) => {
-  await Project.findByPk(req.params.id).then(o => res.send(o.dataValues))
+  const projeto = await Project.findByPk(req.params.id)
+
+  if (projeto) {
+    return res.send(projeto)
+  } else {
+    return res.status(400).json({ error: 'Projeto não existe' })
+  }
 }
 
 exports.projects_delete = async (req, res) => {
   const cidadao = await Project.findByPk(req.params.id)
-  await cidadao.destroy()
-  return res.send({
-    mensagem: 'projeto deletado'
-  })
+  if (cidadao) {
+    await cidadao.destroy()
+    return res.send({
+      mensagem: 'projeto deletado'
+    })
+  } else {
+    return res.status(400).json({ error: 'proejeto não existe' })
+  }
 }
